Major update to the service file structure
