+++ src/services/deleteFromObject.js
@@ -1,0 +1,9 @@
+const deleteFromObject = (targetProperty, targetObject) => {
+    const resultObject = {...targetObject};
+
+    delete resultObject[targetProperty];
+
+    return resultObject;
+};
+
+export default deleteFromObject;
