+++ src/services/filterInArray.js
@@ -1,0 +1,13 @@
+const filterInArray = (toBeChecked,toBeFiltered) => {
+    let filteredArray = [];
+
+    toBeChecked.forEach(element => {
+        if(element !== toBeFiltered){
+            filteredArray.push(element);
+        };
+    });
+
+    return filteredArray;
+};
+
+export default filterInArray;
