+++ src/services/filterOverall.js
@@ -1,0 +1,13 @@
+const filterOverall = (ID,type,iteratable) => {
+    let filtered = -1;
+
+    iteratable.forEach(element => {
+        if(element.ID === parseInt(ID) && element.type === type){
+            filtered = element;
+        };
+    });
+
+    return filtered;
+};
+
+export default filterOverall;
