+++ src/services/findInArray.js
@@ -1,0 +1,13 @@
+const findInArray = (toBeChecked,toBeFound) => {
+    let isFound = false;
+
+    toBeChecked.forEach(element => {
+        if(element === toBeFound){
+            isFound = true;
+        };
+    });
+
+    return isFound;
+};
+
+export default findInArray;
