+++ src/services/getCurrentPageURI.js
@@ -1,0 +1,7 @@
+const getCurrentPageURI = () => {
+    const currentPageURI = window.location.hash.substr(2);
+
+    return currentPageURI;
+};
+
+export default getCurrentPageURI;
