+++ src/services/getCurrentPageUser.js
@@ -1,0 +1,7 @@
+const getCurrentPageUser = (URIPathname) => {
+    const splitted = URIPathname.split('/');
+
+    return splitted[2];
+};
+
+export default getCurrentPageUser;
